backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.4

# Cloud Storage
cloudinary==1.36.0

# Email
resend==2.21.0

# NLP for tag extraction
scikit-learn==1.4.0
nltk==3.8.1

# Security Scanning
python-nmap==0.7.1
requests==2.31.0

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Supabase Integration (updated for compatibility)
supabase==2.10.0
httpx==0.27.2

# AI Content Generation
google-generativeai==0.5.4

# Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
